backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
supabase>=2.3.0
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.26.0

# Date/Time
python-dateutil>=2.8.2

# Background Tasks & Caching
celery>=5.3.0
redis>=5.0.0
celery[redis]>=5.3.0

# MFA / Security
pyotp>=2.9.0
qrcode>=7.4.0
Pillow>=10.0.0

# WebSockets
websockets>=12.0

# Payments (Phase 3)
stripe>=7.0.0

# Communications (Phase 3)
twilio>=8.0.0
sendgrid>=6.9.0

# Google Integration (Phase 3)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Microsoft Integration (Phase 3)
msal>=1.24.0

# ML/Analytics (Phase 4)
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# AI/NLP (Phase 4)
openai>=1.3.0

# Push Notifications (Phase 5)
firebase-admin>=6.2.0
pywebpush>=1.14.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
